task2/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
google-generativeai==0.3.2
python-dotenv==1.0.0
